--- app/src/main/java/com/example/cadeapp/LoginAdapter.java
@@ -1,31 +1,0 @@
-package com.example.cadeapp;
-
-
-import androidx.annotation.NonNull;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.lifecycle.Lifecycle;
-import androidx.viewpager2.adapter.FragmentStateAdapter;
-
-public class LoginAdapter extends FragmentStateAdapter {
-
-
-    public LoginAdapter(@NonNull FragmentManager fragmentManager, @NonNull Lifecycle lifecycle) {
-        super(fragmentManager, lifecycle);
-    }
-
-    @NonNull
-    @Override
-    public Fragment createFragment(int position) {
-
-        if(position == 1){
-            return new RegistroFragment();
-        }
-        return new LoginFragment();
-    }
-
-    @Override
-    public int getItemCount() {
-        return 2;
-    }
-}
