+++ app/src/main/res/drawable/btn_bienvenida1.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <stroke android:color="@color/black" android:width="3dp"/>
+    <corners android:radius="20dp"/>
+
+</shape>
