+++ app/src/main/res/layout/activity_recuperar.xml
@@ -7,80 +7,93 @@
     android:layout_height="match_parent"
     tools:context=".RecuperarActivity">
 
-
-  <ImageView
-      android:id="@+id/imageView"
-      android:layout_width="wrap_content"
-      android:layout_height="0dp"
-      android:scaleType="centerCrop"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintHeight_percent=".25"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintVertical_bias="0" />
-
-  <TextView
-      android:id="@+id/text_titulo"
-      android:layout_width="0dp"
-      android:layout_height="wrap_content"
-      android:textSize="30dp"
-      android:textStyle="bold"
-      app:layout_constraintVertical_bias=".3"
-      app:layout_constraintHorizontal_bias=".5"
-      android:text="Ayuda con la contraseña"
-      app:layout_constraintWidth_percent=".8"
+  <ScrollView
+      android:layout_width="match_parent"
+      android:layout_height="match_parent"
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent" />
+      app:layout_constraintTop_toTopOf="parent">
 
-  <TextView
-      android:id="@+id/text_instr"
-      android:layout_width="0dp"
-      android:layout_height="wrap_content"
-      android:textStyle="bold"
-      android:text="Escribe tu dirección de correo electrónico asociado a tu cuenta de Delicio."
-      app:layout_constraintVertical_bias=".1"
-      app:layout_constraintHorizontal_bias=".5"
-      app:layout_constraintWidth_percent=".8"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/text_titulo" />
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
 
-  <EditText
-      android:id="@+id/correo_recup"
-      android:layout_width="0dp"
-      android:layout_height="wrap_content"
-      android:background="@drawable/edittext_bg"
-      android:hint="Correo"
-      android:paddingLeft="20dp"
-      android:paddingTop="15dp"
-      android:paddingBottom="15dp"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintHorizontal_bias=".5"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/text_instr"
-      app:layout_constraintVertical_bias="0.03"
-      app:layout_constraintWidth_percent=".8" />
 
-  <Button
-      android:id="@+id/recup_password"
-      android:layout_width="0dp"
-      android:layout_height="wrap_content"
-      android:background="@drawable/btn_background"
-      android:text="Recuperar contraseña"
-      android:textColor="@color/white"
-      android:textSize="16sp"
-      android:textStyle="bold"
-      app:layout_constraintVertical_bias="0.3"
-      app:layout_constraintWidth_percent=".8"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/correo_recup" />
+      <LinearLayout
+          android:id="@+id/linearLayout"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:orientation="vertical"
+          android:padding="20dp"
+          android:gravity="center"
+          app:layout_constraintBottom_toBottomOf="parent"
+          app:layout_constraintEnd_toEndOf="parent"
+          app:layout_constraintHorizontal_bias="0.0"
+          app:layout_constraintStart_toStartOf="parent"
+          app:layout_constraintTop_toTopOf="parent">
 
+        <TextView
+            android:id="@+id/textView5"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/ayuda_con_la_contrase_a"
+            android:layout_marginTop="150dp"
+            android:textSize="32sp"
+            android:textStyle="bold" />
+
+        <View
+            android:id="@+id/view1"
+            android:layout_width="match_parent"
+            android:layout_height="30dp" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textAlignment="center"
+            android:text="@string/escribe_tu_direccion"/>
+
+        <View
+            android:id="@+id/view2"
+            android:layout_width="match_parent"
+            android:layout_height="30dp" />
+
+        <com.google.android.material.textfield.TextInputLayout
+            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="15dp"
+            android:hint="@string/correoUsuario"
+            app:endIconMode="clear_text"
+            app:startIconDrawable="@drawable/emailicon">
+
+          <com.google.android.material.textfield.TextInputEditText
+              android:id="@+id/correo_recup"
+              android:layout_width="match_parent"
+              android:layout_height="wrap_content" />
+
+        </com.google.android.material.textfield.TextInputLayout>
+
+
+        <View
+            android:id="@+id/view3"
+            android:layout_width="match_parent"
+            android:layout_height="30dp" />
+
+        <Button
+            android:id="@+id/recup_password"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="@drawable/btn_background"
+            android:text="@string/recuperar_contrase_a"
+            android:textSize="20sp"
+            android:textStyle="bold" />
+
+
+      </LinearLayout>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+  </ScrollView>
 
   </androidx.constraintlayout.widget.ConstraintLayout>
