+++ app/src/main/res/values/colors.xml
@@ -9,4 +9,5 @@
     <color name="blue2">#F21E0080</color>
     <color name="grey">#D3D3D3</color>
     <color name="backopcion">#FFFFFF</color>
+    <color name="red">#FF0000</color>
 </resources>
