+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,31 @@
 <resources>
     <string name="app_name">CadeApp</string>
     <string name="default_web_client_id">639190160106-1hr4mfsk2gcnuhssknjcp3m0fpsj60ua.apps.googleusercontent.com</string>
+
+
+    <!-- String de la pantalla login -->
+    <string name="correoUsuario">Correo electrónico</string>
+    <string name="passwordUsuario">Contraseña</string>
+    <string name="nombreUsuario">Nombre</string>
+    <string name="telefonoUsuario">Teléfono</string>
+    <string name="confirmarpasswordUsuario">Confirmar Contraseña</string>
+    <string name="iniciar">Iniciar sesión</string>
+    <string name="btniniciar">iniciar_sesion</string>
+    <string name="Olvidarpassword">¿Olvidó su contraseña?</string>
+    <string name="iniciarwith">O inicia sesion con</string>
+    <string name="gmail">gmail</string>
+    <string name="bienvenido">Bienvenido</string>
+    <string name="crear_cuenta">Crear Cuenta</string>
+    <string name="tengo_una_cuenta">Tengo una cuenta</string>
+    <string name="por_favor_ingresa_tus_datos">Por Favor ingresa tus datos</string>
+    <string name="registrar_datos">Registrar datos</string>
+    <string name="por_favor_registra_tus_dats">Por favor registra tus dats</string>
+    <string name="tel_fono">Teléfono</string>
+    <string name="nombre">Nombre</string>
+    <string name="confirmar_contrase_a">Confirmar contraseña</string>
+    <string name="ayuda_con_la_contrase_a">Ayuda con la contraseña</string>
+    <string name="escribe_tu_direccion">Escribe tu dirección de correo electrónico asociado a tu cuenta de Delicio.</string>
+    <string name="recuperar_contrase_a">Recuperar contraseña</string>
+
+
 </resources>
