+++ app/src/main/res/values/themes.xml
@@ -1,10 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Base.Theme.Winestastic" parent="Theme.AppCompat.Light.NoActionBar">
+    <style name="Theme.Winestastic" parent="Theme.MaterialComponents.DayNight.NoActionBar.Bridge">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+        <item name="colorPrimary">#4CAF50</item>
         <item name="colorPrimaryDark">#853455</item>
     </style>
-
-    <style name="Theme.Winestastic" parent="Base.Theme.Winestastic" />
 </resources>
