+++ build.gradle.kts
@@ -1,6 +1,6 @@
 buildscript {
     dependencies {
-        classpath("com.google.gms:google-services:4.4.0")
+        classpath("com.google.gms:google-services:4.4.1")
     }
 }
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
